chatbot/requirements.txt
streamlit
openai
python-dotenv
pandas
scikit-learn
tqdm
requests
numpy
tensorflow
tensorflow-hub
speechrecognition
joblib
